+++ telegram/helpers.go
@@ -128,6 +128,9 @@ func (u *Update) EffectiveUser() *User {
 	if u.PurchasedPaidMedia != nil {
 		return &u.PurchasedPaidMedia.From
 	}
+	if u.ManagedBot != nil {
+		return &u.ManagedBot.User
+	}
 	if u.ChatJoinRequest != nil {
 		return &u.ChatJoinRequest.From
 	}